requirements.txt
# Flask & Web
Flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# Google Cloud & AI
google-cloud-aiplatform==1.38.1
google-cloud-storage==2.14.0
google-generativeai==0.3.2
langchain==0.1.0
langchain-google-vertexai==0.0.6

# Vector Database
chromadb==0.4.18

# Database (PostgreSQL)
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
pandas==2.1.4

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# NLP & Embeddings
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.3
joblib==1.3.2

# Environment & Configuration
python-dotenv==1.0.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2
